fix and add specs
